chore: reorganize swizzled files
