+++ docusaurus.config.js
@@ -320,6 +320,7 @@ module.exports = {
       title: navbarTitle,
     },
     metadata: [{ name: 'twitter:card', content: 'summary_large_image' }],
+    hideableSidebar: true,
   },
   themes: [
     [
