+++ src/theme/Layout/index.js
@@ -2,15 +2,16 @@ import React, { useEffect, useState } from 'react';
 import OriginalLayout from '@theme-original/Layout';
 import { ToastContainer } from 'react-toastify';
 
-import URLS from '../../urls';
-import { AuthContextProvider } from '../contexts/AuthContext';
+import URLS from '../../../urls';
+import { AuthContextProvider } from '../../contexts/AuthContext';
 
 const isBrowser = typeof window !== `undefined`;
 const isProd = process.env.DEPLOYMENT_ENV === 'production';
 
 export default function Layout(props) {
   const isRootPath = isBrowser && window.location.pathname === URLS.docs.root;
-  const shouldRedirect = isBrowser && window.location.pathname !== '/video/docs/'
+  const shouldRedirect =
+    isBrowser && window.location.pathname !== '/video/docs/';
   const [canRender, setCanRender] = useState(!isProd || !isRootPath);
 
   // whick redirect for home page. this should happen here in order to avoid
