+++ src/theme/Root/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 
 import OriginalRoot from '@theme-original/Root';
 
-import { DocusaurusContextProvider } from '../contexts/DocusaurusContext';
-import { BreadcrumbsContextProvider } from '../contexts/BreadcrumbsContext';
+import { DocusaurusContextProvider } from '../../contexts/DocusaurusContext';
+import { BreadcrumbsContextProvider } from '../../contexts/BreadcrumbsContext';
 
 export default function DocItem(props) {
   return (
